+++ app/assets/javascripts/channels/evaluations/manage.js
@@ -2,26 +2,32 @@ import consumer from '../consumer';
 
 consumer.subscriptions.create('EvaluationChannel', {
   received(evaluation) {
-    const displayedState = {processing: '実行中', completed: '完了', error: 'エラー'};
-
     const trId = `#${evaluation.evaluation_id}`;
-    if ($(trId).length) {
+    const column = $(trId);
+
+    if (column.length) {
       switch (evaluation.state) {
         case 'processing':
-          this.changeRowColor(trId, evaluation.state);
+          column.removeClass('cursor-auto');
+          column.addClass('table-warning cursor-pointer');
+          column.attr('data-state', evaluation.state);
+          column.attr('title', '結果を確認');
           $(`${trId} > td[class*=performed_at]`).text(evaluation.performed_at);
+          $(`${trId} > td[class*=state]`).text('0%完了');
           break;
         case 'completed':
-          this.changeRowColor(trId, evaluation.state);
-          const column = $(`${trId} > td[class*=state] button`);
-          column.removeClass('btn-warning');
-          column.addClass('btn-success');
-          $(`${trId} > td[class*=state]`).text(displayedState[evaluation.state]);
+          column.removeClass('table-warning');
+          column.addClass('table-success');
+          column.attr('data-state', evaluation.state);
+          column.attr('title', '結果を確認');
+          $(`${trId} > td[class*=state]`).text('完了');
           this.addDownloadButton(trId, evaluation);
           break;
         case 'error':
-          this.changeRowColor(trId, evaluation.state);
-          $(`${trId} > td[class*=state]`).text(displayedState[evaluation.state]);
+          column.removeClass('table-warning cursor-pointer');
+          column.addClass('table-danger cursor-auto');
+          column.attr('data-state', evaluation.state);
+          $(`${trId} > td[class*=state]`).text('エラー');
           break;
         default:
           this.updateProgress(trId, evaluation);
@@ -32,13 +38,6 @@ consumer.subscriptions.create('EvaluationChannel', {
         dataType: 'script',
       });
     }
-  },
-
-  changeRowColor(trId, state) {
-    const stateToClassMap = {processing: 'warning', completed: 'success', error: 'danger'};
-    const column = $(trId);
-    column.removeClass('warning');
-    column.addClass(stateToClassMap[state]);
   },
 
   addDownloadButton(trId, evaluation) {